clean up, submit on enter, componentize splash
